worlds/dk64/requirements.txt
pillow>=12.1.0
